[WIP] Properly decode git utf-8 output on Windows
All git commands which print file names and similar encoding in utf-8 on
Windows.  The encoding of commit messages is controlled by
i18n.logOutputEncoding, but it's easier to set that to UTF-8 than try to
distinguish between commands that respect and those that don't.

magit-log-output-coding-system should probably be obsoleted/removed.